Extract date helpers into library and add tests
