+++ sitegen/Cargo.toml
@@ -11,6 +11,9 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 clap = { version = "4", features = ["derive"] }
 
+[dev-dependencies]
+tempfile = "3"
+
 [[bin]]
 name = "templategen"
 path = "src/templategen.rs"
