+++ sitegen/src/main.rs
@@ -2,6 +2,7 @@ use chrono::{Datelike, NaiveDate, Utc};
 use clap::{Parser, Subcommand};
 use pulldown_cmark::{Options, Parser as CmarkParser, html::push_html};
 use serde::Deserialize;
+use sitegen::read_inline_start;
 use std::collections::BTreeMap;
 use std::fs;
 use std::path::Path;
@@ -20,74 +21,6 @@ enum Commands {
     Validate,
     /// Generate PDFs and HTML into /dist
     Generate,
-}
-
-fn month_from_en(name: &str) -> Option<u32> {
-    match name {
-        "January" => Some(1),
-        "February" => Some(2),
-        "March" => Some(3),
-        "April" => Some(4),
-        "May" => Some(5),
-        "June" => Some(6),
-        "July" => Some(7),
-        "August" => Some(8),
-        "September" => Some(9),
-        "October" => Some(10),
-        "November" => Some(11),
-        "December" => Some(12),
-        _ => None,
-    }
-}
-
-fn month_from_ru(name: &str) -> Option<u32> {
-    match name {
-        "Январь" => Some(1),
-        "Февраль" => Some(2),
-        "Март" => Some(3),
-        "Апрель" => Some(4),
-        "Май" => Some(5),
-        "Июнь" => Some(6),
-        "Июль" => Some(7),
-        "Август" => Some(8),
-        "Сентябрь" => Some(9),
-        "Октябрь" => Some(10),
-        "Ноябрь" => Some(11),
-        "Декабрь" => Some(12),
-        _ => None,
-    }
-}
-
-fn read_inline_start() -> Option<(i32, u32)> {
-    let content = std::fs::read_to_string("cv.md").ok()?;
-    for line in content.lines() {
-        if let Some((month_str, year_str)) = line
-            .trim()
-            .strip_prefix('*')
-            .and_then(|s| s.split_once('–'))
-            .or_else(|| {
-                line.trim()
-                    .strip_prefix('*')
-                    .and_then(|s| s.split_once('—'))
-            })
-        {
-            let year_str = year_str.trim();
-            if year_str.starts_with("Present") || year_str.starts_with("Настоящее время")
-            {
-                let parts: Vec<&str> = month_str.trim().split_whitespace().collect();
-                if parts.len() == 2 {
-                    let (month_text, year_text) = (parts[0], parts[1]);
-                    let year: i32 = year_text.parse().ok()?;
-                    if let Some(month) =
-                        month_from_en(month_text).or_else(|| month_from_ru(month_text))
-                    {
-                        return Some((year, month));
-                    }
-                }
-            }
-        }
-    }
-    None
 }
 
 fn format_duration_en(total_months: i32) -> String {
